+++ internal/mode/usage/markdown/spec_file_test.go
@@ -1,0 +1,65 @@
+package markdown_test
+
+import (
+	"github.com/dogmatiq/ferrite"
+	. "github.com/dogmatiq/ferrite/internal/mode/usage/markdown"
+	"github.com/dogmatiq/ferrite/variable"
+	. "github.com/onsi/ginkgo/v2"
+)
+
+var _ = DescribeTable(
+	"file spec",
+	tableTest(
+		"spec/file",
+		WithoutExplanatoryText(),
+		WithoutIndex(),
+		WithoutUsageExamples(),
+	),
+	Entry(
+		"deprecated",
+		"deprecated.md",
+		func(reg *variable.Registry) {
+			ferrite.
+				File("PRIVATE_KEY", "path to the private key file").
+				Deprecated(ferrite.WithRegistry(reg))
+		},
+	),
+	Entry(
+		"optional",
+		"optional.md",
+		func(reg *variable.Registry) {
+			ferrite.
+				File("PRIVATE_KEY", "path to the private key file").
+				Optional(ferrite.WithRegistry(reg))
+		},
+	),
+	Entry(
+		"required",
+		"required.md",
+		func(reg *variable.Registry) {
+			ferrite.
+				File("PRIVATE_KEY", "path to the private key file").
+				Required(ferrite.WithRegistry(reg))
+		},
+	),
+	Entry(
+		"optional with default value",
+		"with-default.md",
+		func(reg *variable.Registry) {
+			ferrite.
+				File("PRIVATE_KEY", "path to the private key file").
+				WithDefault("/etc/ssh/id_rsa").
+				Optional(ferrite.WithRegistry(reg))
+		},
+	),
+	Entry(
+		"required with default value",
+		"with-default.md",
+		func(reg *variable.Registry) {
+			ferrite.
+				File("PRIVATE_KEY", "path to the private key file").
+				WithDefault("/etc/ssh/id_rsa").
+				Required(ferrite.WithRegistry(reg))
+		},
+	),
+)